==> 01-data-label-encoding.ipynb <==
### Label Encode Categorical Data

import pandas as pd
import numpy as np
from sklearn import preprocessing as prep

# explore and handle data

train_df = pd.read_csv('./data/train.csv')
test_df = pd.read_csv('./data/test.csv')

test_df.head()

train_df.head()

# add test df a survived cloumn

test_df['Survived'] = 0

test_df.head()

# concat train and test sets

concat = train_df.append(test_df, ignore_index=True)

concat.head()

print(train_df.shape)
print(test_df.shape)
print(concat.shape)

train_df.Embarked[881:]

test_df.Embarked[:10]

concat.Embarked[881:901]

concat.Survived[881:901]

concat.Sex.head(10)

concat.Sex.unique()

le = prep.LabelEncoder()

le.fit(concat.Sex)

le.classes_

Sex_le = le.transform(concat.Sex)

Sex_le[0:10]

concat_le = concat.copy()

concat_le.head()

concat_le.Sex = Sex_le

concat_le.head()

concat_le.dtypes

print(concat.Survived.unique())
print(concat.Pclass.unique())
print(concat.Sex.unique())
print(concat.SibSp.unique())
print(concat.Parch.unique())
print(concat.Embarked.unique())

embarked = concat['Embarked'].fillna('0')

embarked.unique()

le.fit(embarked)

le.classes_

embarked = le.transform(embarked)

embarked[:10]

concat_le.Embarked = embarked

concat_le.head(10)

train_le = concat_le.iloc[:891].copy()
test_le = concat_le.iloc[891:].copy()

train_le[880:]

test_le[:10]

train_le.to_csv('./data/train_le.csv', index=False)
test_le.to_csv('./data/test_le.csv', index=False)

%ls data
