==> 02-data-feature-engineering.ipynb <==
### Feature Engineer dataset

This notebook requires  `train_le.csv` and `test_le.csv` datasets (requires running [dataset-01-label-encoding](./data-01-label-encoding.ipynb) first).

import pandas as pd
import numpy as np
from sklearn import preprocessing as prep

train = pd.read_csv('./data/train_le.csv')
test = pd.read_csv('./data/test_le.csv')

train.head()

test.head()

# concat dfs
concat = train.append(test)

concat.shape

train.shape[0] + test.shape[0]

Handle names first

NameSplit = concat.Name.str.split('[,.]')

NameSplit.head()

NameSplit.values[0]

NameSplit.values[:5]

titles = [str.strip(name[1]) for name in NameSplit.values]

titles[:10]

concat['Title'] = titles

concat.head(10)

concat.Title.unique()

# redundancy: combine Mademoiselle and Madame into a single type
concat.Title.values[concat.Title.isin(['Mme', 'Mlle'])] = 'Mlle'

# keep reducing the number of factor levels
concat.Title.values[concat.Title.isin(['Capt', 'Don', 'Major', 'Sir'])] = 'Sir'
concat.Title.values[concat.Title.isin(['Dona', 'Lady', 'the Countess', 'Jonkheer'])] = 'Lady'

le = prep.LabelEncoder()

le.fit(concat.Title)

le.classes_

concat.Title = le.transform(concat.Title)

concat.head(10)

New feature family size

# new feature family size
concat['FamilySize'] = concat.SibSp.values + concat.Parch.values + 1

concat.head(10)

New feature `FamilyID`, extract family information from surnames and family size information. Members of a family should have both the same surname and family size.

surnames = [str.strip(name[0]) for name in NameSplit.values]

surnames[:10]

concat['Surname'] = surnames

concat['FamilyID'] = concat.Surname.str.cat(concat.FamilySize.astype(str), sep='')

concat.head(10)

# mark any family id as small if family size is less than or equal to 2
concat.FamilyID.values[concat.FamilySize.values <= 2] = 'Small'

concat.head(10)

concat.FamilyID.value_counts()

Too many family ids with few family members, maybe some families had different last names or something else. Let's clean this too.

freq = list(dict(zip(concat.FamilyID.value_counts().index.tolist(), concat.FamilyID.value_counts().values)).items())
type(freq)

freq = np.array(freq)
freq[:10]

freq.shape

# select the family ids with frequency of 2 or less
freq[freq[:,1].astype(int) <= 2].shape

freq = freq[freq[:,1].astype(int) <= 2]

freq[:,0]

concat.FamilyID.values[concat.FamilyID.isin(freq[:,0])] = 'Small'

concat.FamilyID.value_counts()

concat.head(10)

# label encoding for family id
le.fit(concat.FamilyID)
concat.FamilyID = le.transform(concat.FamilyID)
concat.FamilyID.unique()

concat_reduce = concat[[
    'PassengerId', 'Pclass', 'Sex', 'Age', 'SibSp',
    'Parch', 'Fare', 'Title', 'Embarked', 'FamilySize',
    'FamilyID', 'Survived']]
concat_reduce.head()

train_final = concat_reduce.iloc[:891].copy()
test_final = concat_reduce.iloc[891:].copy()

test_final.head()

train_final.to_csv('./data/train_final.csv', index=False)
test_final.to_csv('./data/test_final.csv', index=False)

%ls data
