==> 03-split-train-valid.ipynb <==
# Split to Train and Validation sets

import numpy as np
import pandas as pd

train = pd.read_csv('./data/train_final.csv')

train.shape

train.head(10)

# shuffle dataframe
train = train.sample(frac=1).reset_index(drop=True)

train.head(10)

n_valid_rows = int(train.shape[0]*0.3)
n_valid_rows

valid_split = train.iloc[:n_valid_rows].copy()
train_split = train.iloc[n_valid_rows:].copy()
print(train_split.shape)
print(valid_split.shape)

valid_split.head(10)

train_split.head(10)

train_split.to_csv('./data/train_split_final.csv', index=False)
valid_split.to_csv('./data/valid_split_final.csv', index=False)

%ls data
