==> 04-dnn.ipynb <==
# Deep Neural Networks with TensorFlow's Dataset API and Estimators

With TensorFlow 1.4, the Dataset API is [introduced](https://developers.googleblog.com/2017/09/introducing-tensorflow-datasets.html). The real advantage of the Dataset API is that a lot of memory management is done for the user when using large file-based datasets. And, in this work, we will be implementing a predefined DNN estimator and feed it with the Dataset API for Kaggle's Titanic dataset.

---

With Dataset API we can use file-based datasets or datasets in the memory. In this work we will read the data from a csv file. In order to have it, you should first run [this file](./01-data-label-encoding.ipynb) to label encode and then [this file](./02-data-feature-engineering.ipynb) for feature engineering and finally run [this file](split-train-valid.ipynb) to split the train set into train and valid sets.

import pandas as pd
import numpy as np
import tensorflow as tf

Our dataset includes the features:

train = pd.read_csv('./data/train_split_final.csv')
valid = pd.read_csv('./data/valid_split_final.csv')

train.head()

`{Pclass, Sex, Age, SibSp, Parch, Fare, Title, Embarked, FamilySize, FamilyID}` we will be using as the features and the `Survived` column will be our labels.

In order to use the Datasets API and feed the Estimator, we should write an input function like this:

```python
def input_fn():
    ...<code>...
    return ({ 'Pclass':[values], ..<etc>.., 'FamilyID':[values] },
            [Survived])
```

This function takes the `file_path` as input and outputs a two-element tuple. The first element of the tuple is a dictionary containing feature names as keys and features as values. And the second element is a list of labels for the training batch.

Other two arguments for the input function are `perform_shuffle` and `repeat_count`. If `perform_shuffle` is `True` the order of the examples are shuffled. The `perform_shuffle` argument specifies the number of epochs during training, for instance, if `perform_shuffle=1` all the train set examples are passed only once.

And the implementation is as follows, we will use this function to feed the estimator later.

# define feature names first

feature_names = [
    'Pclass',
    'Sex',
    'Age',
    'SibSp',
    'Parch',
    'Fare',
    'Title',
    'Embarked',
    'FamilySize'
    'FamilyID']

def titanic_input_fn(file_path, perform_shuffle=False, repeat_count=1):
    def decode_csv(line):
        # second argument of decode_csv defines the data types for each dataset column!
        # the first argument is passenger ids thus integer
        # the last column is survived or not labels thus integer
        # and the rest are float.
        parsed_line = tf.decode_csv(
            line, [[0], [0.], [0.], [0.], [0.], [0.], [0.], [0.], [0.], [0.], [0.], [0]])
        label = parsed_line[-1:] # Last element is the label
        del parsed_line[-1] # Delete last element (it is the labels)
        features = parsed_line[1:] # First element is excluded since it is the id column
        d = dict(zip(feature_names, features)), label
        return d
    
    dataset = (tf.data.TextLineDataset(file_path) # Read text file
        .skip(1) # Skip header row
        .map(decode_csv)) # Transform each elem by applying decode_csv fn
    if perform_shuffle:
        # Randomizes input using a window of 256 elements (read into memory)
        dataset = dataset.shuffle(buffer_size=256)
    dataset = dataset.repeat(repeat_count) # Repeats dataset this # times
    dataset = dataset.batch(32)  # Batch size to use
    iterator = dataset.make_one_shot_iterator()
    batch_features, batch_labels = iterator.get_next()
    return batch_features, batch_labels

Memory management is provided here with `TextLineDataset`.

Let's print and check the first batch:

train_path = './data/train_final.csv'

next_batch = titanic_input_fn(train_path, False) # Will return first 32 elements

with tf.Session() as sess:
    first_batch = sess.run(next_batch)
print(first_batch)

Ok, it is working!

---
Now we will define our DNN estimator

%rm -r ./checkpoints

# path to save checkpoints
save_dir = './checkpoints'

# reset default graph if rebuilding the classifier
tf.reset_default_graph()

# Create the feature_columns, which specifies the input to our model.
# All our input features are numeric, so use numeric_column for each one.
feature_columns = [tf.feature_column.numeric_column(k) for k in feature_names]

# define the classifier
classifier = tf.estimator.DNNClassifier(
    feature_columns=feature_columns, # The input features to our model
    hidden_units=[2048, 1024, 512, 256, 128], # 5 layers
    n_classes=2, # survived or not {1, 0}
    model_dir=save_dir, # Path to where checkpoints etc are stored
    optimizer=tf.train.RMSPropOptimizer(
        learning_rate=0.00001),
    dropout=0.1)

Now we will train the model using `titanic_input_fn` and our classifier.

train_path = './data/train_split_final.csv'
valid_path = './data/valid_split_final.csv'
test_path = './data/test_final.csv'

# the classifier will run for 500 epochs below
classifier.train(input_fn=lambda: titanic_input_fn(train_path, True, 500))

# evaluate
# Return value will contain evaluation_metrics such as: loss & average_loss
evaluate_result = classifier.evaluate(
   input_fn=lambda: titanic_input_fn(valid_path, False, 1))
print('')
print("Evaluation results:")
for key in evaluate_result:
    print("   {}, was: {}".format(key, evaluate_result[key]))

Predict on test set

predict_results = classifier.predict(
    input_fn=lambda: titanic_input_fn(test_path, False, 1))
print("Predictions on test file")
i=892
for prediction in predict_results:
    # Will print the predicted class: 0 or 1.
    print(prediction["class_ids"][0])
    i += 1
print(i)

That's all folks!
